Refactor query handler.
